Add type parameter to /files service for jsonp submission.
Problems can now be loaded with the type=jsonp parameter and the codecheck report will load in the same page with an ajax and jsonp call.

When the parameter type=jsonp, two script tags are added to the page. One is jQuery, and the other is the ajax/jsonp submission handler when the problem form is submitted.