Add receiveMessage method for iframe message passing.
A score from a codecheck problem is sent in the following format:

    {
        "correct": int,
        "errors": 0,
        "maxscore": int,
        "repo": string,
        "problem": string
    }

For example,

    {"correct":2,"errors":0,"maxscore":5,"repo":"bj4cc","problem":"ch06/c06_exp_6_105"}

There's no current way to keep track of "errors" from a codecheck
problem, but it's kept there to be consistent with the interactive
exercises messages. It will always be 0.

Codecheck exercises have a "repo" and "problem" key instead of an
"activity" key from the interactive exercises.